move x updater to a new c++ file
